Refactor get schemas.
