+++ src/scripts/helpers.js
@@ -12,6 +12,7 @@ import {
     transactionInProgressShow,
     transactionSuccess
 } from "./store.js";
+import {VAULT_INFORMATION_QUERY} from "./queries.js";
 
 
 export async function getEventArgs(tx, eventName, contract) {
@@ -345,7 +346,8 @@ export async function showPrompt(transaction, options) {
     promptSuccessText.set("Transaction successful!")
     promptNoBottom.set(false)
     promptBottomText.set("")
-    promptCloseAction.set(()=>{})
+    promptCloseAction.set(() => {
+    })
     transactionHash.set("false")
 
     //show prompt
@@ -391,7 +393,8 @@ export async function showPromptSFTCreate(transaction, options) {
     promptSuccessText.set("Transaction successful!")
     promptNoBottom.set(false)
     promptBottomText.set("")
-    promptCloseAction.set(()=>{})
+    promptCloseAction.set(() => {
+    })
     transactionHash.set("false")
 
     //show prompt
@@ -454,4 +457,63 @@ export function mapOrder(array, order, key) {
     });
 
     return array;
+}
+
+export async function getSchemas(activeNetwork, vault, deposits) {
+    let tempSchema = []
+
+
+    let variables = {id: vault.address.toLowerCase()}
+    if (vault.address) {
+
+        try {
+            let resp = await getSubgraphData(activeNetwork, variables, VAULT_INFORMATION_QUERY, 'offchainAssetReceiptVault')
+            let receiptVaultInformations = []
+
+            if (resp && resp.data && resp.data.offchainAssetReceiptVault) {
+                receiptVaultInformations = resp.data.offchainAssetReceiptVault.receiptVaultInformations
+
+                if (receiptVaultInformations.length) {
+                    for (let i = 0; i < receiptVaultInformations.length; i++) {
+                        let cborDecodedInformation = cborDecode(receiptVaultInformations[i].information.slice(18))
+                        let schemaHash = cborDecodedInformation[1].get(0)
+                        let url = `${IPFS_GETWAY}${schemaHash}`
+                        let assetCount = getAssetCount(schemaHash, deposits)
+
+                        try {
+                            if (url) {
+                                let res = await axios.get(url)
+                                if (res) {
+                                    tempSchema = [...tempSchema, {
+                                        ...res.data,
+                                        timestamp: receiptVaultInformations[i].timestamp,
+                                        id: receiptVaultInformations[i].id,
+                                        hash: schemaHash,
+                                        assetCount,
+                                    }]
+                                    tempSchema = tempSchema.filter(d => d.displayName)
+                                }
+                            }
+                        } catch (err) {
+                            console.log(err)
+                        }
+                    }
+                    return tempSchema
+                }
+            }
+
+        } catch (err) {
+            console.log(err)
+        }
+    }
+}
+
+function getAssetCount(hash, deposits) {
+    let depositsWithSchema = deposits.filter(d => d.receipt.receiptInformations[0]?.schema === hash)
+    let depositAmounts = depositsWithSchema.map(d => d.amount);
+    let assetCount = depositAmounts.reduce(
+        (accumulator, currentValue) => ethers.BigNumber.from(accumulator).add(ethers.BigNumber.from(currentValue)),
+        0
+    );
+    return ethers.utils.formatUnits(assetCount, 18)
 }
